Move color Block and BlockEntity code to separate classes
