+++ src/main/java/nl/dyonb/karam/common/block/entity/ElevatorBlockEntity.java
@@ -11,46 +11,8 @@ import net.minecraft.nbt.CompoundTag;
 import nl.dyonb.karam.common.block.ElevatorBlock;
 import nl.dyonb.karam.registry.KaramBlockEntityTypes;
 
-public class ElevatorBlockEntity extends BlockEntity implements BlockEntityClientSerializable {
-    private int color = 16777215;
-
-    public int getColor() { return color; }
-
-    public void setColor(int colorToSet) {
-        this.color = colorToSet;
-        markDirty();
-
-        if (this.world != null && !this.world.isClient) {
-            sync();
-        }
-    }
-
+public class ElevatorBlockEntity extends BaseColorBlockEntity {
     public ElevatorBlockEntity() {
         super(KaramBlockEntityTypes.ELEVATOR);
-    }
-
-    @Override
-    public void fromTag(BlockState state, CompoundTag tag) {
-        super.fromTag(state, tag);
-        fromClientTag(tag);
-    }
-
-    @Override
-    public CompoundTag toTag(CompoundTag tag) {
-        super.toTag(tag);
-        return toClientTag(tag);
-    }
-
-    @Override
-    public void fromClientTag(CompoundTag tag) {
-        color = tag.getInt("color");
-    }
-
-    @Override
-    public CompoundTag toClientTag(CompoundTag tag) {
-        if (color == 0)
-            color = 16777215;
-        tag.putInt("color", color);
-        return tag;
     }
 }
